% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalFunctions.R
\name{getProdData}
\alias{getProdData}
\title{Animal Functions}
\usage{
getProdData(animal, func, area, year)
}
\arguments{
\item{animal}{character of animal as listed in \code{\link{codeTable}}.}

\item{func}{character. Whether for protein or energy function - All of the
form {animal}_{energy|protein}_factor}

\item{area}{character. M49 area codes}

\item{year}{character. Years}
}
\description{
Functions mostly used in animal functions. These are for retrieving data from
the SWS.
}

