% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animalCoefficients}
\alias{animalCoefficients}
\title{Coefficients for intensity factor calculations for each animal group}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 8 rows and 2 columns.}
\usage{
animalCoefficients
}
\description{
Coefficients for intensity factor calculations for each animal group
}
\keyword{datasets}

