% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animalCPCGroup}
\alias{animalCPCGroup}
\title{Animal Species CPC codes and corresponding animal group}
\format{a data.table with 14 rows and 2 columns
\itemize{
 \item measuredItemCPC, key, CPC code for animal species
 \item animalGroup, Animal Group 
 }}
\usage{
animalCPCGroup
}
\description{
The table links livestock species to animal groups which share the same 
characteristics in husbandry. Intensity factors are derived for each animal
group.
}
\keyword{datasets}

