% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAnimalUnits.R
\name{calculateAnimalUnits}
\alias{calculateAnimalUnits}
\title{Calculate protein and energy demand of individual animals}
\usage{
calculateAnimalUnits()
}
\description{
The main output of this function is a data.table of Animal Unit Indices
(AUIs)
}
\section{Animals}{

The module currently has equations for the following animals:

\itemize{
\item 1. Cattle
\item 2. Buffalo
\item 3. Sheep
\item 4. Goats
\item 5. Camels
\item 6. Pigs
\item 7. Chickens
\item 8. Ducks
\item 9. Geese
\item 10. Turkeys
\item 11. Horses
\item 12. Rabbits
}
}

