% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIR.R
\name{calculateIR}
\alias{calculateIR}
\title{Calculate intensity factor for a given country in a given year}
\usage{
calculateIR(density_replacements = "1")
}
\arguments{
\item{density_replacements}{Groups which use livestock density instead of agricultural productivity.}
}
\description{
This function pulls in data from GLEAM to do so
}

