% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeTable}
\alias{codeTable}
\title{Code table for animal functions}
\format{a data.table with 54 rows and 8 columns
\itemize{
 \item module, Name of animal and animal function
 \item fun, Demand estimation type (energy or protein)
 \item table, Dataset from which data are to be retrieved
 \item measuredItemCPC, key, CPC code for animal species, meat or milk
 \item CPCDescription, Description of CPC element
 \item measuredElement, key, Code for element
 \item elementDescription, Name of element
 \item variable, Variable name used in animal requrirements functions
 }}
\usage{
codeTable
}
\description{
Different codes and classifications for animals and products derived from them (meat and milk).
}
\keyword{datasets}

