% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/faoswsFeed-package.R
\docType{package}
\name{faoswsFeed-package}
\alias{faoswsFeed}
\alias{faoswsFeed-package}
\title{Support package for the Feed module}
\description{
This package contains most of the functions needed for the Feed module to run.
}
\author{
Bernhard Dalheimer, Sebastian Campbell, Onno Hoffmeister
}

