% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQueryKey.R
\name{getQueryKey}
\alias{getQueryKey}
\title{Get Query Key}
\usage{
getQueryKey(dimension, datasetkey = swsContext.datasets[[1]])
}
\arguments{
\item{dimension}{character. Name of dimension you wish to extract from
DatasetKey}

\item{datasetkey}{Datasetkey. DatasetKey from which you wish to extract
dimensions}
}
\description{
Get a data key from the query
}

