% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYearCodes.R
\name{getYearCodes}
\alias{getYearCodes}
\title{Get years corresponding to the source coverage}
\usage{
getYearCodes(source = c("faostat", "trade_module"), allyears = NULL)
}
\arguments{
\item{source}{character.}

\item{allyears}{character.}
}
\value{
character. Years that belong to both the source set and allyears, if provided
}
\description{
In November 2017, a request was made to use old trade data for 2010-2013 and
new trade data for 2014-2015... On the face of it, this is the wrong way to
go about the problem. It's messy and bad practice. A parameter should be made
to switch between the two and it should be run as two separate modules.
BUT... The years are not all entirely independent. The gap between 2013-2014
is tricky because there are animal functions that need to look at the year
after the the current one. As a result, it's best to make a function able to
combine these cases.
}

