% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landData}
\alias{landData}
\title{Land data taken from FAOSTAT regarding meadows and pastures}
\format{a data.table with 5160 rows and 3 columns
#' \itemize{
 \item geographicAreaM49, key, M49 code for country
 \item landElement, Area (1000 ha) which is code 5110
 \item itemLand, meadows and pastures item code 6655
 \item timepointYears, key, Year
 \item permanentMeadows, Value
}}
\usage{
landData
}
\description{
This data is a set taken from FAOSTAT 
(\url{http://faostat3.fao.org/download/R/RL/E}) with all countries for the 
years 1990-2013. It contains area (ha, code: 5110) of meadows and pastures 
(code: 6655). This dataset contains imputed values which were calculated 
based on lending group in the World Bank country classification table 
('class.csv', 
\url{http://data.worldbank.org/about/country-and-lending-groups}). For the 
calculation, small islands are excluded from the imputation model. This is 
defined by 'FaostatAreaGroupLists.csv'
(\url{http://faostat.fao.org/site/371/default.aspx}).
}
\keyword{datasets}

