% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimizeFeed.R
\name{optimizeFeed}
\alias{optimizeFeed}
\title{Optimise feed}
\usage{
optimizeFeed(x, y)
}
\description{
This function optimizes the amount of feed (in metric tonnes) of each commodity such that
energy and protein requiremetns are both fulfilled using a linear program
}

