% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s2005}
\alias{s2005}
\title{Survey data for intensification of animal husbandry}
\format{a data.table with 1848 rows and 4 columns
\itemize{
 \item geographicAreaM49, key, M49 code for country
 \item animalGroup, key, Animal Group
 \item timepointYears, key, Year
 \item measuredItemCPC, key, CPC code for items
 \item intensity, intensity factor for 2005
}}
\usage{
s2005
}
\description{
The table is derived from feed basket data on country level which has been collected in the framework
of the Global Livestock Environmental Assessment Model (GLEAM) for the year of 2005 and is not publicly available.
For more information:
(\url{http://www.fao.org/gleam/en/})
}
\keyword{datasets}

