% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stockCodes}
\alias{stockCodes}
\title{Animal names and Codes}
\format{a data.table with 14 rows and 2 columns
\itemize{
 \item module, key, name of animal species 
 \item measuredItemCPC, key, CPC code for animal species
}}
\usage{
stockCodes
}
\description{
The table links animal names included in the model to CPC codes.
}
\keyword{datasets}

