% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wdi}
\alias{wdi}
\title{Labor productivity in agriculture (World Bank)}
\format{a data.table with 7208 rows and 3 columns
\itemize{
 \item geographicAreaM49, key, M49 code for country
 \item timepointYears, key, Year
 \item productivity, Labour productivty in agriculture as value added per
 worker in constant 2005 US$
}}
\usage{
wdi
}
\description{
The indicator is part of the World Development Incicators (WDI), published by
the World Bank and describes value added per worker. It's based on a
downloaded flat csv. All empty columns were removed.
(\url{http://data.worldbank.org/indicator/EA.PRD.AGRI.KD})
}
\keyword{datasets}

