% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animalCoefficients}
\alias{animalCoefficients}
\title{Coefficients for intensity factor calculations}
\format{a data.table with 54 rows and 8 columns
\itemize{
 \item animalGroup, key, Animal group
 \item coefficient, Coefficient 
}}
\usage{
animalCoefficients
}
\description{
Coefficients which are used in the estimation equations of livestock
intensity of each animal group.
}
\keyword{datasets}

