% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codeTable}
\alias{codeTable}
\title{Code table for animal functions}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 54 rows and 8 columns.}
\usage{
codeTable
}
\description{
Code table for animal functions
}
\keyword{datasets}

