% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeFeed.R
\name{optimizeFeed}
\alias{optimizeFeed}
\title{Optimise feed}
\usage{
optimizeFeed(x, y)
}
\arguments{
\item{x}{character. M49 area code, length 1.}

\item{y}{character. Year, length 1.}
}
\description{
This function optimizes the amount of feed (in metric tonnes) of each commodity such that 
energy and protein requirements are both fulfilled using a linear program.
}

